Create Splash.py
Designed to let user know system is up and happy in a headless scenario. Uses auto_write=True for both the effect, and to see how quickly a command, or more importantly batch of commands) can be sent to the panel. This is designed for an 8x32 panel cut down to 8x18 for specific purpose.